Working on main, fixed bug with p2 as IterativeBot
Had it accessing p1, but by the time we got to p2, old value p1 was
already written over. As a result, made a new global variable
(p1OldMove) and made p1Move a normal local variable. This is not an
issue with p2, because it is always played second.